requirements.txt
anyio==3.6.2
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.9.24
ffmpeg-python==0.2.0
ffprobe==0.5
future==0.18.2
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==3.4
python-telegram-bot==20.0a6
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tornado==6.1
tzdata==2022.7
tzlocal==4.2
